requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.12
pymongo
certifi
scikit-learn
